# docs/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: SIC API
  version: 0.1.0
  description: Implementation of accounting system based in hordak and following SIC
    (Andrisani) definitions.
  contact:
    name: gastong256
paths:
...
  /api/v1/companies/{company_id}/journal/:
    get:
      operationId: list_journal_entries
      description: Returns paginated journal entries for the company ordered by entry
        number.
      summary: List journal entries
      parameters:
      - in: path
        name: company_id
        schema:
          type: integer
        required: true
      - in: query
        name: page
        schema:
          type: integer
        description: Pagination page number.
      tags:
      - journal
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntryListPaginated'
          description: ''
        '409':
          description: Company must be opened before journal is available
        '401':
          description: Authentication required
        '403':
          description: Not the company owner or teacher
        '404':
          description: Company not found
    post:
      operationId: create_journal_entry
      description: Creates a new journal entry and immediately posts it. Debits must
        equal credits. All accounts must belong to this company.
      summary: Create and post a journal entry
      parameters:
      - in: path
        name: company_id
        schema:
          type: integer
        required: true
      tags:
      - journal
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JournalEntryCreate'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntryDetail'
          description: ''
        '400':
          description: Validation or business rule error
        '409':
          description: Company not opened, closed period, read-only demo company,
            or concurrent conflict
        '401':
          description: Authentication required
        '403':
          description: Not the company owner or teacher
        '404':
          description: Company not found
  /api/v1/companies/{company_id}/journal/{entry_id}/:
    get:
      operationId: get_journal_entry
      description: Returns full detail of a journal entry including all lines.
      summary: Retrieve a journal entry
      parameters:
      - in: path
        name: company_id
        schema:
          type: integer
        required: true
      - in: path
        name: entry_id
        schema:
          type: integer
        required: true
      tags:
      - journal
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntryDetail'
          description: ''
        '409':
          description: Company must be opened before journal is available
        '401':
          description: Authentication required
        '403':
          description: Not the company owner or teacher
        '404':
          description: Company or entry not found
  /api/v1/companies/{company_id}/journal/{entry_id}/reverse/:
    post:
      operationId: reverse_journal_entry
      description: Creates a full contra-entry (debits<->credits) and links it to
        the original entry.
      summary: Reverse a journal entry
      parameters:
      - in: path
        name: company_id
        schema:
          type: integer
        required: true
      - in: path
        name: entry_id
        schema:
          type: integer
        required: true
      tags:
      - journal
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JournalEntryReverse'
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JournalEntryDetail'
          description: ''
        '400':
          description: Validation error
        '401':
          description: Authentication required
        '403':
          description: Not the company owner or teacher
        '404':
          description: Company or entry not found
        '409':
          description: Company not opened, already reversed, closed period, or read-only
            demo company
...
  /api/v1/companies/{id}/:
    get:
      operationId: get_company
      summary: Retrieve a company
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this company.
        required: true
      tags:
      - companies
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
          description: ''
    put:
      operationId: update_company
      summary: Update a company
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this company.
        required: true
      tags:
      - companies
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompanyWrite'
        required: true
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
          description: ''
        '400':
          description: Validation error
        '403':
          description: Permission denied
        '409':
          description: Read-only demo company
    patch:
      operationId: partial_update_company
      summary: Partially update a company
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this company.
        required: true
      tags:
      - companies
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCompanyWrite'
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Company'
          description: ''
        '400':
          description: Validation error
        '403':
          description: Permission denied
        '409':
          description: Read-only demo company
    delete:
      operationId: delete_company
      summary: Delete a company
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this company.
        required: true
      tags:
      - companies
      security:
      - jwtAuth: []
      - cookieAuth: []
      responses:
        '204':
          description: No content
        '403':
          description: Permission denied
        '409':
          description: Company has protected accounting records
...
components:
  schemas:
...
    Company:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          maxLength: 255
        description:
          type: string
          description: Optional business description for a more realistic simulation.
        tax_id:
          type: string
          description: Simulated CUIT (optional).
          maxLength: 20
        owner_username:
          type: string
          readOnly: true
        account_count:
          type: integer
          readOnly: true
        books_closed_until:
          type: string
          format: date
          readOnly: true
          nullable: true
          description: Entries on or before this date are locked.
        is_demo:
          type: boolean
          readOnly: true
          description: Marks a seeded demo company shared across the system.
        is_read_only:
          type: boolean
          readOnly: true
          description: If true, accounting writes and metadata updates are blocked.
        has_opening_entry:
          type: boolean
          readOnly: true
        accounting_ready:
          type: boolean
          readOnly: true
        opening_entry_id:
          type: integer
          nullable: true
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - account_count
      - accounting_ready
      - books_closed_until
      - created_at
      - has_opening_entry
      - id
      - is_demo
      - is_read_only
      - name
      - opening_entry_id
      - owner_username
      - updated_at
...
    CompanyWrite:
      type: object
      properties:
        name:
          type: string
          description: Company name.
          maxLength: 255
        description:
          type: string
          default: ''
          description: Optional business description.
        tax_id:
          type: string
          default: ''
          description: Simulated CUIT (optional).
          maxLength: 20
      required:
      - name
...
    JournalEntryCreate:
      type: object
      properties:
        date:
          type: string
          format: date
          description: Accounting date of the entry.
        description:
          type: string
          description: Glosa / description of the entry.
          maxLength: 500
        source_type:
          allOf:
          - $ref: '#/components/schemas/SourceTypeEnum'
          default: MANUAL
        source_ref:
          type: string
          default: ''
          description: E.g. "Factura A-0001".
          maxLength: 100
        lines:
          type: array
          items:
            $ref: '#/components/schemas/JournalEntryLineWrite'
          description: At least one DEBIT and one CREDIT line, balanced.
      required:
      - date
      - description
      - lines
    JournalEntryDetail:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        entry_number:
          type: integer
          readOnly: true
          title: Número de asiento
          description: Correlativo por empresa, asignado automáticamente.
        date:
          type: string
          format: date
          readOnly: true
          title: Fecha
        description:
          type: string
          readOnly: true
          title: Descripción
        source_type:
          allOf:
          - $ref: '#/components/schemas/SourceTypeEnum'
          readOnly: true
          title: Tipo de comprobante
        source_ref:
          type: string
          readOnly: true
          title: Referencia
          description: 'Ej.: "Factura A-0001", "Recibo 00123".'
        created_by:
          type: string
          readOnly: true
        reversal_of_id:
          type: integer
          readOnly: true
        reversed_by_id:
          type: integer
          nullable: true
          readOnly: true
        total_debit:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          readOnly: true
        total_credit:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          readOnly: true
        lines:
          type: array
          items:
            $ref: '#/components/schemas/JournalEntryLineRead'
          readOnly: true
      required:
      - created_by
      - date
      - description
      - entry_number
      - id
      - lines
      - reversal_of_id
      - reversed_by_id
      - source_ref
      - source_type
      - total_credit
      - total_debit
    JournalEntryLineRead:
      type: object
      properties:
        account_id:
          type: integer
          title: Cuenta
          readOnly: true
        account_code:
          type: string
          readOnly: true
        account_name:
          type: string
          readOnly: true
        type:
          allOf:
          - $ref: '#/components/schemas/TypeEnum'
          readOnly: true
          title: Tipo
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          readOnly: true
          title: Importe
      required:
      - account_code
      - account_id
      - account_name
      - amount
      - type
    JournalEntryLineWrite:
      type: object
      properties:
        account_id:
          type: integer
          description: ID of the level-3 account.
        type:
          allOf:
          - $ref: '#/components/schemas/TypeEnum'
          description: |-
            DEBIT or CREDIT.

            * `DEBIT` - Deudora
            * `CREDIT` - Acreedora
        amount:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          description: Positive amount.
      required:
      - account_id
      - amount
      - type
    JournalEntryList:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        entry_number:
          type: integer
          readOnly: true
          title: Número de asiento
          description: Correlativo por empresa, asignado automáticamente.
        date:
          type: string
          format: date
          readOnly: true
          title: Fecha
        description:
          type: string
          readOnly: true
          title: Descripción
        source_type:
          allOf:
          - $ref: '#/components/schemas/SourceTypeEnum'
          readOnly: true
          title: Tipo de comprobante
        source_ref:
          type: string
          readOnly: true
          title: Referencia
          description: 'Ej.: "Factura A-0001", "Recibo 00123".'
        created_by:
          type: string
          readOnly: true
        reversal_of_id:
          type: integer
          readOnly: true
        reversed_by_id:
          type: integer
          nullable: true
          readOnly: true
        total_debit:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          readOnly: true
        total_credit:
          type: string
          format: decimal
          pattern: ^-?\d{0,13}(?:\.\d{0,2})?$
          readOnly: true
      required:
      - created_by
      - date
      - description
      - entry_number
      - id
      - reversal_of_id
      - reversed_by_id
      - source_ref
      - source_type
      - total_credit
      - total_debit
    JournalEntryListPaginated:
      type: object
      properties:
        count:
          type: integer
        next:
          type: string
          nullable: true
        previous:
          type: string
          nullable: true
        results:
          type: array
          items:
            $ref: '#/components/schemas/JournalEntryList'
      required:
      - count
      - next
      - previous
      - results
    JournalEntryReverse:
      type: object
      properties:
        date:
          type: string
          format: date
          description: Date for the reversal entry.
        description:
          type: string
          description: Optional description for the reversal entry.
          maxLength: 500
...
    PatchedCompanyWrite:
      type: object
      properties:
        name:
          type: string
          description: Company name.
          maxLength: 255
        description:
          type: string
          default: ''
          description: Optional business description.
        tax_id:
          type: string
          default: ''
          description: Simulated CUIT (optional).
          maxLength: 20
...
    SourceTypeEnum:
      enum:
      - OPENING
      - MANUAL
      - INVOICE
      - RECEIPT
      - OTHER
      type: string
      description: |-
        * `OPENING` - Apertura
        * `MANUAL` - Manual
        * `INVOICE` - Factura
        * `RECEIPT` - Recibo
        * `OTHER` - Otro
...
    TypeEnum:
      enum:
      - DEBIT
      - CREDIT
      type: string
      description: |-
        * `DEBIT` - Deudora
        * `CREDIT` - Acreedora
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sessionid
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT